requirements.txt
numpy
torch>=1.2
torchvision
matplotlib
tensorflow
tensorboard
terminaltables
pillow
tqdm
